Create Convert file to csv or any suitable format
# Download the dataset source file [GSE137829 ] from the supplementary section,
decompress it and extract the files to your working directory.
#Locate any one ".txt" file that contains the gene expression data. This file
contains the raw counts of gene expression.
#Read the gene expression data from the chosen ".txt" file using a programming
environment suitable for data manipulation (e.g., Python, R).
# Create a data frame from the parsed data, setting the gene names as row indices
and the cell IDs or barcodes as column names. Export this data to a .csv file,
containing genes or features listed in rows and cell IDs or barcodes in columns.